get fases por categoria

+++ models/CargaOPs.py
@@ -131,7 +131,7 @@ def OPemProcesso(empresa, AREA, filtro = '-'):
 
         consulta = pd.merge(consulta,leadTime2,on=['codFase','categoria'], how='left')
         consulta.fillna('-', inplace=True)
-
+        consulta['meta2'] = consulta['meta2'].astype(int)
 
         consulta['meta'] = consulta.apply(lambda row : row['meta'] if row['meta2'] == '-' else row['meta2'], axis=1)
         consulta.drop('meta2', axis=1, inplace=True)
